Work around problem with a composite build that includes Spring Boot
Prior to this change, a composite build that included Spring Boot
would fail when trying to resolve dependencies of
spring-boot-gradle-plugin. For reasons that are not yet understood,
the dependency constraints in Boot's build were not applied so the
dependencies were unversioned, making them unresolvable.

This commit works around the problem by disabling dependency
substitution for the buildscript classpath configuration. This is
an improvement, but it's not a solution to the problem as it means
that a change to spring-boot-gradle-plugin cannot be tested via an
included build. However, it should now be possible to test changes
to all of Spring Boot's main modules.