Fix task finalization
It would appear that when a task is a finalizer for a task that is being
run is a finalizer it does not run. Unfortunately, this mean that
compose down would not run after the app had been stopped following a
test failure.

This commit simplifies the task finalization configuration so that a
finalizing tasks are no longer chained. Instead, the compose down task
is now a finalizer for the compose up task and the stop app task is a
finalizer for the  start app task. Task dependencies and ordering
(mustRunAfter) are used to ensure that the tasks run in the required
order:

1. compose up
2. start app
3. run tests
4. stop app
5. compose down

These changes should also ensure that compose down runs in the event of
start app failing.

Closes gh-6