Make the smoke tests compatible with docker-compose v2
Unlike v1, docker-compose v2 does not allow capital letters to be
used in project names. The Docker Compose Gradle plugin derives the
project name from the name of the ComposeSettings. By removing
capital letters from the name of the settings, the project name can
be made compatible with docker-compose v2.