Split tests into four separate tasks
There are now four separate tasks that run tests:

- appTest

This is the task that was previously named jvmAotTest. It starts the
application on the JVM using java -jar and runs tests found in the
appTest source set against it.

- nativeAppTest

This is the task that was previously named nativeAotTest. It starts the
application as a native executable and runs tests found in the appTest
source set against it.

- test

This is the standard test task that runs the tests found in the test
source set. The task is customized to use AOT-generated artifacts by
settings the spring.aot.enabled system property to true.

- nativeTest

This is the standard task contributed by native image plugin. It
executes the native test binary using the tests found in the test
source set following AOT processing.

CI has been updated so that the four tasks are run separately. This
should provide a clearer picture of what does and does not work.

Closes gh-121