Improve diagnostics for test initialization failures
Previously, the application's log output was only reported through
an exception's message. This resulted in it appearing in the local
HTML test report but not in the information captured in a build
scan. This impedes diagnosis of CI failures where the HTML test
report isn't readily available.

This commit updates AwaitApplication to write the message to
System.err as well. This results in it being captured in the build
scan allowing it to be view in Gradle Enterprise.