Add a task to warm Gradle's caches
A new warmCaches task is added to each smoke test. This task downloads
all of the dependencies on the classpaths of the test's main, test,
and appTest source sets. It is intended for us in GitHub Actions to
make more efficient use of its cache and increase the likelihood of
a cache hit when running each individual set of smoke tests.