Try to improve GitHub Actions cache efficiency
Previously, every individual smoke test would write to and read from
the GitHub Actions cache. This quickly exhausted to 10GB limit,
resulting in cache misses and time spent downloading Gradle and
dependencies.

This commit configures each smoke test's workflow to only read from
the GitHub Actions cache and to never write to it. To populate the
cache, it introduces a new Warm Caches workflow that has read-write
access to the cache. It runs once per branch per day before the
branch's smoke tests run. The intent is that the Warm Caches workflow
will populate the GitHub Actions cache and that each of the
individual smoke tests will then reuse this cache. With each branch
only have a single workflow that writes to the cache, the 10GB limit
should not be exhausted, increasing cache hits.