Restore use of Spring Boot Starter Data Redis
This commit restores the use of the Spring Data Redis starter that was
removed as part as moving to the Spring Boot 4.x structure. The Spring
Boot module for Spring Session Redis has an implementation dependency on
Spring Boot Data Redis. As a result, it would not end up on the compile
classpath. However, as this sample does use Redis directly, it should
arguably have a direct dependency to it as well.