Add profile smoke test
This commit adds a smoke test exercising the retention of a profile
enabled during AOT processing. It validates the profile has been enabled
with profile-specific configuration as well as enabling it when running
on the JVM and in a native image.

In particular, this makes sure that profile-specific configuration files
are loaded as expected.

Closes gh-338